+++ packages/back-end/src/services/eventForwarderDatasourceLifecycle.ts
@@ -1,7 +1,7 @@
 import { DataSourceInterface } from "shared/types/datasource";
 import { EventForwarderConfigInterface } from "shared/validators";
+import { getEventForwarderSinkTypeForDatasource } from "shared/util";
 import { usingFileConfig } from "back-end/src/init/config";
-import { getEventForwarderSinkTypeForDatasource } from "back-end/src/services/eventForwarderConfig";
 import { teardownEventForwarderInfrastructureRemote } from "back-end/src/services/eventForwarderProvisioning";
 import { logger } from "back-end/src/util/logger";
 import { ApiReqContext } from "back-end/types/api";
