--- packages/shared/src/util/event-forwarder-feature-usage-queries.ts
@@ -1,101 +1,0 @@
-import type { FeatureUsageQuery } from "shared/types/datasource";
-import {
-  buildBigQueryEventForwarderTableReference,
-  buildSnowflakeEventForwarderTableReference,
-  EVENT_FORWARDER_AVRO_PARTITION_FIELD,
-} from "./event-forwarder-fact-table";
-import { normalizeSnowflakeTableNameForEventForwarder } from "./snowflake-table-name";
-
-export const EVENT_FORWARDER_FEATURE_USAGE_TABLE = "feature_usage";
-
-export type BuildEventForwarderFeatureUsageTableRefParams =
-  | {
-      sinkType: "bigquery";
-      projectId: string;
-      dataset: string;
-    }
-  | {
-      sinkType: "snowflake";
-      database: string;
-      schema: string;
-    };
-
-export function buildEventForwarderFeatureUsageTableReference(
-  params: BuildEventForwarderFeatureUsageTableRefParams,
-): string {
-  if (params.sinkType === "bigquery") {
-    return buildBigQueryEventForwarderTableReference(
-      params.projectId,
-      params.dataset,
-      EVENT_FORWARDER_FEATURE_USAGE_TABLE,
-    );
-  }
-
-  return buildSnowflakeEventForwarderTableReference(
-    params.database,
-    params.schema,
-    normalizeSnowflakeTableNameForEventForwarder(
-      EVENT_FORWARDER_FEATURE_USAGE_TABLE,
-    ),
-  );
-}
-
-export function buildEventForwarderFeatureUsageQuerySql({
-  sinkType,
-  tableRef,
-}: {
-  sinkType: "bigquery" | "snowflake";
-  tableRef: string;
-}): string {
-  if (sinkType === "bigquery") {
-    return `SELECT
-  timestamp AS timestamp,
-  feature_key AS feature_key
-FROM ${tableRef}
-WHERE ${EVENT_FORWARDER_AVRO_PARTITION_FIELD} BETWEEN '{{startDate}}' AND '{{endDate}}'`;
-  }
-
-  return `SELECT
-  TIMESTAMP AS timestamp,
-  FEATURE_KEY AS feature_key
-FROM ${tableRef}`;
-}
-
-export type GenerateEventForwarderFeatureUsageQueryParams =
-  BuildEventForwarderFeatureUsageTableRefParams;
-
-export function buildEventForwarderFeatureUsageQuery(
-  params: GenerateEventForwarderFeatureUsageQueryParams,
-): Pick<FeatureUsageQuery, "query" | "managedBy"> {
-  const tableRef = buildEventForwarderFeatureUsageTableReference(params);
-
-  return {
-    managedBy: "api",
-    query: buildEventForwarderFeatureUsageQuerySql({
-      sinkType: params.sinkType,
-      tableRef,
-    }),
-  };
-}
-
-export function isEventForwarderManagedFeatureUsageQuery(
-  query: FeatureUsageQuery,
-): boolean {
-  return query.managedBy === "api";
-}
-
-export function getActiveFeatureUsageQuery(
-  queries: FeatureUsageQuery[] | undefined,
-): FeatureUsageQuery | undefined {
-  if (!queries?.length) {
-    return undefined;
-  }
-
-  return queries.find(isEventForwarderManagedFeatureUsageQuery) ?? queries[0];
-}
-
-export function eventForwarderManagedFeatureUsageQueryExists(
-  queries: FeatureUsageQuery[],
-): boolean {
-  return queries.some(isEventForwarderManagedFeatureUsageQuery);
-}
