+++ packages/shared/test/util/event-forwarder-destination.test.ts
@@ -1,13 +1,122 @@
 import {
+  DEFAULT_EVENT_FORWARDER_BIGQUERY_TABLE_NAME,
+  DEFAULT_EVENT_FORWARDER_SNOWFLAKE_TABLE_NAME,
   formatBigQueryEventForwarderDestination,
   formatSnowflakeEventForwarderDestination,
+  isValidBigQueryTableName,
+  isValidSnowflakeTableName,
+  normalizeBigQueryTableNameForEventForwarder,
+  normalizeSnowflakeEventForwarderAccessUrl,
+  normalizeSnowflakeTableNameForEventForwarder,
   parseBigQueryEventForwarderDestination,
   parseSnowflakeEventForwarderDestination,
-} from "../../src/util/event-forwarder-destination";
-import {
-  normalizeSnowflakeEventForwarderAccessUrl,
   tryDeriveSnowflakeAccessUrlFromAccount,
-} from "../../src/util/event-forwarder-snowflake-url";
+} from "../../src/util/event-forwarder-destination";
+
+describe("normalizeBigQueryTableNameForEventForwarder", () => {
+  it("returns default when raw is empty", () => {
+    expect(normalizeBigQueryTableNameForEventForwarder("")).toBe(
+      DEFAULT_EVENT_FORWARDER_BIGQUERY_TABLE_NAME,
+    );
+    expect(normalizeBigQueryTableNameForEventForwarder("   ")).toBe(
+      DEFAULT_EVENT_FORWARDER_BIGQUERY_TABLE_NAME,
+    );
+  });
+
+  it("maps hyphens and spaces to underscores", () => {
+    expect(normalizeBigQueryTableNameForEventForwarder("gb-events")).toBe(
+      "gb_events",
+    );
+    expect(normalizeBigQueryTableNameForEventForwarder("gb events")).toBe(
+      "gb_events",
+    );
+  });
+
+  it("prefixes when the first character would be a digit", () => {
+    expect(normalizeBigQueryTableNameForEventForwarder("42foo")).toBe("_42foo");
+  });
+
+  it("preserves unicode letters", () => {
+    expect(normalizeBigQueryTableNameForEventForwarder("événements")).toBe(
+      "événements",
+    );
+  });
+
+  it("throws when there are no letters or digits", () => {
+    expect(() => normalizeBigQueryTableNameForEventForwarder("___")).toThrow(
+      /letter or number/,
+    );
+    expect(() => normalizeBigQueryTableNameForEventForwarder("---")).toThrow(
+      /letter or number/,
+    );
+  });
+});
+
+describe("isValidBigQueryTableName", () => {
+  it("accepts unicode letters and underscores", () => {
+    expect(isValidBigQueryTableName("gb_events")).toBe(true);
+    expect(isValidBigQueryTableName("événements")).toBe(true);
+    expect(isValidBigQueryTableName("_42")).toBe(true);
+  });
+
+  it("rejects hyphens", () => {
+    expect(isValidBigQueryTableName("gb-events")).toBe(false);
+  });
+});
+
+describe("normalizeSnowflakeTableNameForEventForwarder", () => {
+  it("returns default when raw is empty", () => {
+    expect(normalizeSnowflakeTableNameForEventForwarder("")).toBe(
+      DEFAULT_EVENT_FORWARDER_SNOWFLAKE_TABLE_NAME,
+    );
+    expect(normalizeSnowflakeTableNameForEventForwarder("   ")).toBe(
+      DEFAULT_EVENT_FORWARDER_SNOWFLAKE_TABLE_NAME,
+    );
+  });
+
+  it("maps hyphens and spaces to underscores and uppercases", () => {
+    expect(normalizeSnowflakeTableNameForEventForwarder("gb-events")).toBe(
+      "GB_EVENTS",
+    );
+    expect(normalizeSnowflakeTableNameForEventForwarder("gb events")).toBe(
+      "GB_EVENTS",
+    );
+  });
+
+  it("prefixes when the first character would be a digit", () => {
+    expect(normalizeSnowflakeTableNameForEventForwarder("42foo")).toBe(
+      "_42FOO",
+    );
+  });
+
+  it("allows dollar signs", () => {
+    expect(normalizeSnowflakeTableNameForEventForwarder("gb$events")).toBe(
+      "GB$EVENTS",
+    );
+  });
+
+  it("throws when there are no letters or digits", () => {
+    expect(() => normalizeSnowflakeTableNameForEventForwarder("___")).toThrow(
+      /letter or number/,
+    );
+    expect(() => normalizeSnowflakeTableNameForEventForwarder("---")).toThrow(
+      /letter or number/,
+    );
+  });
+});
+
+describe("isValidSnowflakeTableName", () => {
+  it("accepts normalized Snowflake identifiers", () => {
+    expect(isValidSnowflakeTableName("GB_EVENTS")).toBe(true);
+    expect(isValidSnowflakeTableName("_42")).toBe(true);
+    expect(isValidSnowflakeTableName("GB$EVENTS")).toBe(true);
+  });
+
+  it("rejects lowercase and hyphens", () => {
+    expect(isValidSnowflakeTableName("gb_events")).toBe(false);
+    expect(isValidSnowflakeTableName("GB-EVENTS")).toBe(false);
+  });
+});
 
 describe("parseBigQueryEventForwarderDestination", () => {
   it("parses dataset.table", () => {
